requirements.txt

# ADK 核心依赖
google-adk>=1.21.0

# Web 框架
fastapi>=0.115.0,<0.124.0
uvicorn>=0.22.0

# HTTP 请求和 API
requests>=2.31.0
PyGithub>=2.8.0

# Git 操作
GitPython>=3.1.0

# 环境变量管理
python-dotenv>=1.0.0

# YAML 配置
PyYAML>=6.0.0,<7.0.0>

# Google Cloud Firestore
google-cloud-firestore
